fix: 프로필 설정하기 버튼이 보이지 않는 버그 수정
